% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twomeans.r
\name{twomeans}
\alias{twomeans}
\title{Two-sample t-test power calculation}
\usage{
twomeans(m1 = NULL, m2 = NULL, n1 = NULL, n2 = NULL, sd = NULL,
  sd1 = NULL, sd2 = NULL, sig.level = 0.05, power = 0.8)
}
\arguments{
\item{m1}{Mean of a group 1 (e.g., the control-group)}

\item{m2}{Mean of a group 2 (e.g., the experimental-group)}

\item{n1}{Number of obs. in group 1 (e.g., the control-group)}

\item{n2}{Number of obs. in group 2 (e.g., the control-group)}

\item{sd}{Standard deviation of each group, i.e., sd = sd1 = sd2}

\item{sd1}{Standard deviation of a group 1 (e.g., the control-group)}

\item{sd2}{Standard deviation of a group 2 (e.g., the experimental-group)}

\item{sig.level}{significance level; default is sig.level = 0.05}

\item{power}{one minus the probability of type II error; default is power = 0.8}
}
\value{
returns an object with all the study parameters
}
\description{
Two-sample t-test power calculation
}
\examples{
twomeans(m1 = 12, m2 = 16, sd = 5)
}
