% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered.r
\name{clustered}
\alias{clustered}
\title{Power calculations in an experiment with group clusters.}
\usage{
clustered(unclustered, rho, obsclus = NULL, numclus = NULL)
}
\arguments{
\item{unclustered}{Results from twomeans not adjusting for clusters}

\item{rho}{Specifies the intraclass correlation coefficient}

\item{obsclus}{Number of observations in each cluster}

\item{numclus}{Maximum number of clusters}
}
\value{
returns an object with all the study parameters
}
\description{
Power calculations in an experiment with group clusters.
}
\examples{
twomeans(m1 = 12, m2 = 16, sd = 5) \%>\% clustered(obsclus = 10, rho = 0.3)
}

